requirements.txt
python-telegram-bot==21.4
openai==1.45.0
python-dotenv==1.0.1
APScheduler==3.10.4
python-dateutil==2.9.0.post0
